mfs: Add basic metadata operation for MFS
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/release-management/issues/ID5W1P

--------------------------------

It is easy to add metadata operation in MFS. In MFS, metadata is
from the lower layer, so we just need to pass the requests to the
lower inode/dentry and adjust the reference.

In remote mode, the cache layer may be empty, so we should create
the target file when lookup.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>