xcall2.0: Fix module unload issue by fixing refcount
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/release-management/issues/ID5CMS

--------------------------------

After the process exits and xcall is unregistered, kernel module cannot
be unloaded properly.

The reason is that mmput() just need to reduce the refcount
of 'mm->mm_users' to 0 which indicates that no user-space processes
are using this address space, but mmdrop() need to reduce the refcount
of 'mm->mm_count' to zero, which indicates that all references to
the memory descriptor have reached zero and the mm_struct can now
be safely freed and this condition may not be satisfied after process
exit. So put_xcall() in mmdrop() can not guarantee the refcount of xcall
will be decremented immediately, so module_put() in put_xcall() may
not called.

Fix this problem by split clear_xcall_area() into free_xcall_area()
and clear_xcall_area(), call put_xcall() in mmput() and free
the mm->xcall in mmdrop().

Fixes: b05676644e95 ("xcall2.0: Add xcall_area")
Signed-off-by: Jinjie Ruan <ruanjinjie@huawei.com>