xcall2.0: prefetch: epoll_ctl no need to occupy a file refcount
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/release-management/issues/ID5CMS

--------------------------------

My test is restart redis server and send kill command to kill it
repeatedly. Then I found when the second time to start redis, it
cannot run and reported 6379 port is been occpuied. It means 6379
port refcount is not dec to 0.

In epoll_ctl(), it will fget file then fput it in close(). But if
we kill the process, it has no chance to do the close() in prefetch
module. Thus, the 6379 port is not released.

Since we only need the struct file address, actually we don't need
to add the refcount. Just fget then fput both in epoll_ctl().

Signed-off-by: Xinyu Zheng <zhengxinyu6@huawei.com>