xsched: add xsched_{set,get}attr syscall
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ID9IFI

-----------------------------------------

If CONFIG_XCU_CGROUP is not enabled, the xsched kernel defaults to only the
RT scheduling class, and there is no interface available to configure the
RT priority. Therefore, two new syscalls are added to set and get the
configuration of the RT scheduling class.
The default priority for RT is 4 (the lowest priority), and it inherits the
priority configuration of the parent process during fork. If pid=0, it
indicates that the configuration is applied to the current process.
Additionally, an interface is reserved for configuring the xsched
scheduling class, which will be used to support CFS in the future.
Currently, these system calls are only available for ARM64 and x86_64
architectures.

Signed-off-by: Liu Kai <liukai284@huawei.com>