scsi: libsas: do not add end deivces to wide port
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/ID9GQH

--------------------------------

Upstream: No

In commit c49b589b3326 ("scsi: libsas: fix issue of swapping two sas
disks") we described a scenario of swapping disks which will cause disk
lost. But now we found that the hardware may behave much slower than we
think. Even the disks of two slots are swapped, the sas address we read
out may still the same. This means even with the prev fix the end device
may be added to a wideport anyway.

So do not allowed the phy of end device type to be added to a wideport
in sas_ex_join_wide_port(). And we need to clear the sas address so that
it can be discovered in next revalidation.

CC: chenxiang <chenxiang66@hisilicon.com>
Signed-off-by: Jason Yan <yanaijie@huawei.com>

DTS:DTS2019051011917
Description: NA
Team: HISI_SW
Feature or Bugfix: Bugfix

Change-Id: I6b32412cb81f16f2574f3b3fa2c0e1373a36ac65
Signed-off-by: x00470154 <x00470154@huawei.com>
Reviewed-on: http://10.90.31.173:8080/7520
Tested-by: public TuringEE <turingee@huawei.com>
Reviewed-by: chenxiang 00284940 <chenxiang66@hisilicon.com>
Reviewed-by: public TuringEE <turingee@huawei.com>
Signed-off-by: Li Lingfeng <lilingfeng3@huawei.com>